+++ packages/shader-ast-stdlib/src/screen/uv.ts
@@ -2,15 +2,21 @@ import {
     $x,
     $xy,
     $y,
+    add,
     assign,
+    bvec4,
     defn,
     div,
+    greaterThan,
+    lessThan,
     mul,
     ret,
     sym,
+    vec2,
     Vec2Sym,
     Vec2Term,
     Vec4Term,
+    _any,
 } from "@thi.ng/shader-ast";
 import { fit0111 } from "../math/fit";
 
@@ -43,4 +49,30 @@ export const aspectCorrectedUV = defn(
             ret(uv),
         ];
     }
+);
+
+/**
+ * Returns true if at least one coordinate of the given point is within the
+ * `width` internal border region of UV rect ([0,0] .. [1,1]).
+ *
+ * ```c
+ * borderMask(vec2(0.91, 0.5), 0.1) // true
+ * borderMask(vec2(0.2, 0.01), 0.1) // true
+ * borderMask(vec2(0.2, 0.2), 0.1) // false
+ * ```
+ */
+export const borderMask = defn(
+    "bool",
+    "borderMask",
+    ["vec2", "float"],
+    (uv, width) => [
+        ret(
+            _any(
+                bvec4(
+                    lessThan(uv, vec2(width)),
+                    greaterThan(add(uv, width), vec2(1))
+                )
+            )
+        ),
+    ]
 );